Fix Arquitectura
