+++ src/generales/index.ts
@@ -17,4 +17,6 @@ export * from './gateway';
 export * from './zona';
 export * from './evento';
 export * from './sucesos';
-export * from './ambiente';
+export * from './cultivo';
+export * from './cultivo';
+export * from './tipoSuelo';
