+++ src/generales/tipoSuelo/index.ts
